broke tests out into smaller pieces. working on server tests
